+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # Changelog
 
 ## 0.6.0
+  * Add support for Group and Project Members
+  * Update Users with Member info 
   * Fetch additional {'default', 'can_push'} attributes for Branches
   * Fetch additional {'authored_date', 'committed_date', 'parent_ids'} attributes for Commits
   * Fetch additional {'upvotes', 'downvotes', 'merge_requests_count' 'weight'} attributes for Issues